Moved DependencyTracker to UpdateManager
To be used for cleaning up unused backup dependency
